+++ client/src/pages/FindUser.tsx
@@ -1,20 +1,14 @@
-import { Label, TextInput, Toast } from 'flowbite-react';
-import { HiCheck } from 'react-icons/hi';
+import { Label, TextInput } from 'flowbite-react';
 import UserContainer from '../components/findUsers/UserContainer';
+import Toast from '../components/Utils/Toast';
 import { useState } from 'react';
 import { useAppSelector } from '../redux/store';
-import { useAppDispatch } from '../redux/store';
-import { toast } from '../redux/feats/globalSlice/globalSlice';
 const FindUser = () => {
   const [nameSearch, setNameSearch] = useState<string>('');
   const [schoolSearch, setSchoolSearch] = useState<string>('');
   const [majorSearch, setMajorSearch] = useState<string>('');
-
-  const dispatch = useAppDispatch();
   const friendToast = useAppSelector((state: any) => state.globalSlice.toast);
-  const addFriendHandler = () => {
-    dispatch(toast(false));
-  };
+
   return (
     <div className="flex min-h-full flex-col items-center space-y-10 bg-slate-400 p-6 dark:bg-slate-900">
       <div>
@@ -76,19 +70,7 @@ const FindUser = () => {
         nameSearch={nameSearch}
         majorSearch={majorSearch}
       />
-      {friendToast.state && (
-        <div className="fixed bottom-4 left-4 z-50">
-          <Toast className="border shadow-lg">
-            <div className="inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-lg bg-green-500 text-white shadow-md dark:bg-green-800 dark:text-green-200">
-              <HiCheck className="h-5 w-5" />
-            </div>
-            <div className="ml-3 text-sm font-semibold text-gray-800 dark:text-gray-300">
-              {friendToast.message}
-            </div>
-            <Toast.Toggle onClick={addFriendHandler} />
-          </Toast>
-        </div>
-      )}
+      {friendToast.state && <Toast message={friendToast.message} />}
     </div>
   );
 };
